// cityflow-fe/src/app/models/financial-report.ts
export interface FinancialReport {
    id: number;
    dailyTicketSales: number;
    monthlyTicketSales: number;
    annualTicketSales: number;
    sponsorshipRevenue: number;
    totalSalariesPaid: number;
    fuelCosts: number;
    vehicleMaintenanceCosts: number;
    reportDate: string;
    netProfit: number;
  }
  